wip: patching -sec test to work against native edgex
Modified a few files in order to get -sec test to work. Note, this is not setting the environment variable nor have a I ran this test as part of the -all flag. To get a successful result I started a fresh install of both edgexfoundry and app-service-configurable. Next, I executed the importCoreDataDump.sh file followed by run.sh -sec